chore: move add_package to pending_project_changes
